+++ js/settings.js
@@ -280,6 +280,14 @@
                 $("#onlyofficeAddrSave").click();
             }
         });
+
+        $("#onlyofficeSecret-show").click(function () {
+            if ($("#onlyofficeSecret").attr("type") == "password") {
+                $("#onlyofficeSecret").attr("type", "text");
+            } else {
+                $("#onlyofficeSecret").attr("type", "password");
+            }
+        });
     });
 
 })(jQuery, OC);