+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Change Log
 
+## Added
+- hide secret key in settings
+
 ## 6.2.0
 ## Changed
 - the ability to use forcesave for federated share files